Create separate component and move search input from filters to toolbar
